+++ rusqlite/Changelog.md
@@ -8,7 +8,9 @@
     * `SqliteRows` is now `Rows`
     * `SqliteRow` is now `Row`
     * `SqliteOpenFlags` is now `OpenFlags`
-  The old, prefixed names are still exported should be considered deprecated.
+    * `SqliteTransaction` is now `Transaction`.
+    * `SqliteTransactionBehavior` is now `TransactionBehavior`.
+  The old, prefixed names are still exported but are deprecated.
 * Adds a variety of `..._named` methods for executing queries using named placeholder parameters.
 * Adds `backup` feature that exposes SQLite's online backup API.
 * Adds `functions` feature that allows user-defined scalar functions to be added to